+++ apps/web/views/evaluation.py
@@ -2,56 +2,12 @@ import csv
 import io
 from uuid import UUID
 
-import django_rq
 from django.conf import settings
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import Http404, HttpResponse
-from django.shortcuts import render, redirect
 from django.views import View
-from django.views.generic import CreateView
 
-from apps.core.jobs import basic_job
-from apps.core.models import Task, Evaluation
-from apps.web.forms import EvaluationForm
-
-
-class CreateEvaluationView(LoginRequiredMixin, CreateView):
-    model = Evaluation
-    form_class = EvaluationForm
-    template_name = "web/evaluation.html"
-
-    def __init__(self):
-        self.object = None
-
-    def form_valid(self, form):
-        self.object = form.save(commit=False)
-        self.object.save()
-
-        with open(self.object.links.path, "r", newline="") as csvfile:
-            reader = csv.DictReader(csvfile, delimiter=";")
-            for row in reader:
-                task = Task.objects.create(
-                    user=self.request.user,
-                    assigment=self.object.assignment,
-                    executor=Task.Executor.EVALUATION,
-                    image=row["link"],
-                    additional_information=row,
-                )
-                self.object.tasks.add(task)
-                django_rq.enqueue(basic_job, task.pk, False)
-        return redirect("evaluation-overview")
-
-
-class EvaluationOverview(LoginRequiredMixin, View):
-    def get(self, request):
-        if not request.user.is_superuser:
-            raise Http404()
-
-        try:
-            evaluation = Evaluation.objects.all()
-        except Evaluation.DoesNotExist:
-            raise Http404()
-        return render(request, "web/evaluation_overview.html", {"evaluations": evaluation})
+from apps.core.models import Evaluation
 
 
 class EvaluationResult(LoginRequiredMixin, View):
@@ -99,4 +55,4 @@ class EvaluationResult(LoginRequiredMixin, View):
         return response
 
 
-__all__ = ["CreateEvaluationView", "EvaluationOverview", "EvaluationResult"]
+__all__ = ["EvaluationResult"]